Refactor API services and update data models
Refactored the function implementations in proxy, phone and rate services in api folder to improve code readability and to use Promises in place of Observables for handling asynchronous operations. Updated the RateGetSimpleListResponseDTO data model to use the newly added RateGetResponseDTO. Removed unnecessary comments and imports from the service files, clarifying and simplifying the codebase.